requirements.txt
absl-py
altair
annotated-types
asttokens
astunparse
attrs
blinker
cachetools
catboost
certifi
charset-normalizer
click
cloudpickle
colorama
contourpy
cramjam
cycler
dacite
dash
dash-core-components
dash-html-components
dash-table
decorator
executing
fastparquet
Flask
flatbuffers
fonttools
fsspec
gast
gitdb
GitPython
google-pasta
graphviz
grpcio
h5py
htmlmin
idna
ImageHash
imbalanced-learn
imblearn
importlib_metadata
ipython
ipython_pygments_lexers
itsdangerous
jedi
Jinja2
joblib
jsonschema
jsonschema-specifications
keras
kiwisolver
libclang
llvmlite
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
ml-dtypes
multimethod
mypy
mypy-extensions
namex
narwhals
nest-asyncio
networkx
numba
numpy
opt_einsum
optional-django
optree
packaging
pandas
parso
patsy
phik
pillow
plotly
prompt_toolkit
protobuf
psutils
pure_eval
puremagic
pyarrow
pydantic
pydantic_core
pydeck
Pygments
pyparsing
pypdf
python-dateutil
pytz
PyWavelets
PyYAML
react
referencing
requests
retrying
rich
rpds-py
scikit-learn
scipy
seaborn
shap
six
sklearn-compat
slicer
smmap
stack-data
statsmodels
streamlit
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
tensorflow_intel
termcolor
threadpoolctl
toml
tornado
tqdm
traitlets
Transcrypt
typeguard
typing_extensions
tzdata
urllib3
visions
watchdog
wcwidth
Werkzeug
wordcloud
wrapt
xgboost
ydata-profiling
zipp